# openapi.yaml
openapi: 3.1.0
info:
  title: Home Expense API
  description: API for managing home expenses and income
  version: 1.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /v1/incomes:
    get:
      tags:
        - Incomes
      summary: Get list of incomes
      operationId: getIncomes
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: year
          in: query
          description: Filter by year
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Successful response
          headers:
            Link:
              description: Pagination links (RFC 5988)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags:
        - Incomes
      summary: Create a new income
      operationId: createIncome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomePostRequest"
      responses:
        "201":
          description: Income created successfully
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/incomes/{id}:
    get:
      tags:
        - Incomes
      summary: Get income by ID
      operationId: getIncomeById
      parameters:
        - $ref: "#/components/parameters/IncomeId"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeGetResponse"
    put:
      tags:
        - Incomes
      summary: Update an income
      operationId: updateIncome
      parameters:
        - $ref: "#/components/parameters/IncomeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomePutRequest"
      responses:
        "204":
          description: Income updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags:
        - Incomes
      summary: Delete an income
      operationId: deleteIncome
      parameters:
        - $ref: "#/components/parameters/IncomeId"
      responses:
        "204":
          description: Income deleted successfully
  /v1/income-attributes:
    get:
      tags:
        - Income Attributes
      summary: Get list of income attributes
      operationId: getIncomeAttributes
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Successful response
          headers:
            Link:
              description: Pagination links (RFC 5988)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeAttributeListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags:
        - Income Attributes
      summary: Create a new income attribute
      operationId: createIncomeAttribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeAttributePostRequest"
      responses:
        "201":
          description: Income attribute created successfully
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/income-attributes/{id}:
    get:
      tags:
        - Income Attributes
      summary: Get income attribute by ID
      operationId: getIncomeAttributeById
      parameters:
        - $ref: "#/components/parameters/IncomeAttributeId"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeAttributeGetResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Income Attributes
      summary: Update an income attribute
      operationId: updateIncomeAttribute
      parameters:
        - $ref: "#/components/parameters/IncomeAttributeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeAttributePutRequest"
      responses:
        "204":
          description: Income attribute updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Income Attributes
      summary: Delete an income attribute
      operationId: deleteIncomeAttribute
      parameters:
        - $ref: "#/components/parameters/IncomeAttributeId"
      responses:
        "204":
          description: Income attribute deleted successfully
        "409":
          $ref: "#/components/responses/Conflict"
components:
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "about:blank"
            title: "Bad Request"
            status: 400
            detail: "descriptionは必須入力です"
            instance: "/v1/expenses"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "about:blank"
            title: "Not Found"
            status: 404
            detail: "経費が見つかりません"
            instance: "/v1/expenses/123"
    Conflict:
      description: Resource conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "about:blank"
            title: "Conflict"
            status: 409
            detail: "既に登録されています"
            instance: "/v1/expense-attributes"
  parameters:
    Page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    IncomeId:
      name: id
      in: path
      required: true
      description: Income ID
      schema:
        type: string
        format: uuid
    IncomeAttributeId:
      name: id
      in: path
      required: true
      description: Income attribute ID
      schema:
        type: string
        format: uuid
  schemas:
...
    IncomePostRequest:
      type: object
      required:
        - description
        - amount
        - receive_date
        - attribute_id
      properties:
        description:
          type: string
          maxLength: 512
        amount:
          type: integer
          minimum: 1
        receive_date:
          type: string
          format: date
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          example: "2026-04-25"
        attribute_id:
          type: string
          format: uuid
    IncomePutRequest:
      type: object
      required:
        - description
        - amount
        - receive_date
        - attribute_id
      properties:
        description:
          type: string
          maxLength: 512
        amount:
          type: integer
          minimum: 1
        receive_date:
          type: string
          format: date
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          example: "2026-04-25"
        attribute_id:
          type: string
          format: uuid
    IncomeGetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        amount:
          type: integer
        receive_date:
          type: string
          format: date
        income_attribute:
          $ref: "#/components/schemas/IncomeAttributeResponse"
    IncomeListResponse:
      type: object
      properties:
        incomes:
          type: array
          items:
            $ref: "#/components/schemas/IncomeGetResponse"
    IncomeAttributePostRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 512
    IncomeAttributePutRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 512
    IncomeAttributeGetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    IncomeAttributeListResponse:
      type: object
      properties:
        income_attributes:
          type: array
          items:
            $ref: "#/components/schemas/IncomeAttributeGetResponse"
    IncomeAttributeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "about:blank"
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: "Bad Request"
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: "descriptionは必須入力です"
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem
          example: "/v1/expenses"